+++ Cognite.Simulator.Utils/DefaultClasses/DefaultModelLibrary.cs
@@ -1,4 +1,3 @@
-using System;
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
@@ -10,13 +9,26 @@ using Microsoft.Extensions.Logging;
 
 namespace Cognite.Simulator.Utils
 {
-
+    /// <summary>
+    /// A default version of the model library
+    /// Should be used as a base class for anyone implementing their own connector
+    /// </summary>
+    /// <typeparam name="TAutomationConfig"></typeparam>
     public class DefaultModelLibrary<TAutomationConfig> :
     ModelLibraryBase<ModelStateBase, DefaultModelFileStatePoco, ModelParsingInfo>
     where TAutomationConfig : AutomationConfig, new()
     {
         private ISimulatorClient<ModelStateBase, SimulatorRoutineRevision> __simulationClient;
 
+        /// <summary>
+        /// Construct a default model library
+        /// </summary>
+        /// <param name="config"></param>
+        /// <param name="cdf"></param>
+        /// <param name="logger"></param>
+        /// <param name="client"></param>
+        /// <param name="simulationClient"></param>
+        /// <param name="store"></param>
         public DefaultModelLibrary(
             DefaultConfig<TAutomationConfig> config,
             CogniteDestination cdf,
@@ -25,7 +37,7 @@ namespace Cognite.Simulator.Utils
             ISimulatorClient<ModelStateBase, SimulatorRoutineRevision> simulationClient,
             IExtractionStateStore store = null) :
             base(
-                config.Connector.ModelLibrary,
+                config?.Connector.ModelLibrary,
                 new List<SimulatorConfig> { config.Simulator },
                 cdf,
                 logger,
@@ -35,22 +47,33 @@ namespace Cognite.Simulator.Utils
             __simulationClient = simulationClient;
         }
 
+        /// <summary>
+        /// Extract the model information, should be implemented on the instance of the ISimulatorClient
+        /// </summary>
+        /// <param name="state"></param>
+        /// <param name="token"></param>
+        /// <returns></returns>
         protected override async Task ExtractModelInformation(
         ModelStateBase state,
         CancellationToken token)
         {
 
             if (__simulationClient != null) {
-                __simulationClient.ExtractModelInformation(state, token);
+                await __simulationClient.ExtractModelInformation(state, token).ConfigureAwait(false);
             } else {
                 state.CanRead = true;
                 state.ParsingInfo.SetSuccess();
             }
         }
 
+        /// <summary>
+        /// Helps map the local state.db from the API data
+        /// </summary>
+        /// <param name="modelRevision"></param>
+        /// <returns></returns>
         protected override ModelStateBase StateFromModelRevision(SimulatorModelRevision modelRevision)
         {
-            var modelState = new DefaultModelFilestate(modelRevision.Id.ToString())
+            var modelState = new DefaultModelFilestate(modelRevision?.Id.ToString())
             {
                 UpdatedTime = modelRevision.LastUpdatedTime,
                 ModelExternalId = modelRevision.ModelExternalId,
