+++ Sample.Calculator.Simulator/Simulator.cs
@@ -14,13 +14,6 @@ public class CalculatorSimulatorClient :
         _logger = logger;
     }
 
-    public void ExtractModelInformation(ModelStateBase state, CancellationToken _token)
-    {
-        _logger.LogCritical("ExtractModelInformation WAS CALLED IN SIMULATORAUTOMATION");
-        state.CanRead = false;
-        state.ParsingInfo.SetFailure();
-    }
-
     public string GetConnectorVersion()
     {
         return CommonUtils.GetAssemblyVersion();
@@ -53,6 +46,14 @@ public class CalculatorSimulatorClient :
         {
         }
         
+    }
+
+    Task ISimulatorClient<ModelStateBase, SimulatorRoutineRevision>.ExtractModelInformation(ModelStateBase state, CancellationToken _token)
+    {
+        _logger.LogCritical("ExtractModelInformation WAS CALLED IN SIMULATORAUTOMATION");
+        state.CanRead = false;
+        state.ParsingInfo.SetFailure();
+        return Task.CompletedTask;
     }
 }
